+++ hrp/pkg/uixt/video_crawler.go
@@ -1,6 +1,8 @@
 package uixt
 
 import (
+	"math"
+	"math/rand"
 	"time"
 
 	"github.com/pkg/errors"
@@ -43,8 +45,6 @@ type VideoCrawler struct {
 
 	// used to help checking if swipe success
 	failedCount int64
-	lastFeed    *FeedVideo
-	lastLive    *LiveRoom
 
 	FeedCount int            `json:"feed_count"`
 	FeedStat  map[string]int `json:"feed_stat"` // 分类统计 feed 数量：视频/图文/广告/特效/模板/购物
@@ -112,142 +112,15 @@ func (vc *VideoCrawler) isTargetAchieved() bool {
 	return vc.isFeedTargetAchieved() && vc.isLiveTargetAchieved()
 }
 
-func (vc *VideoCrawler) checkLiveVideo(feedVideo *FeedVideo) (enterPoint PointF, yes bool) {
-	// TODO: check if preview-live from feedVideo
-	if feedVideo.Type != "live" {
-		return PointF{}, false
-	}
-
-	// take screenshot and get OCR texts via image service
-	texts, err := vc.driverExt.GetScreenTexts()
-	if err != nil {
-		return PointF{}, false
-	}
-
-	// 预览流入口：DY/KS
-	// 标签文案：点击进入直播间|进入直播间领金币
-	points, err := texts.FindTexts([]string{".*进入直播间.*"}, WithScope(0, 0.3, 1, 0.8), WithRegex(true))
-	if err == nil {
-		return points[0].Center(), true
-	}
-	// 标签文案：直播中|直播卖货|直播团购
-	points, err = texts.FindTexts([]string{"直播中|直播卖货|直播团购"},
-		WithScope(0, 0.7, 0.5, 1), WithRegex(true))
-	if err == nil {
-		return points[0].Center(), true
-	}
-
-	// 预览流入口：KS/KSLite
-	// 评论框文案：和主播聊聊天...|聊聊天...
-	points, err = texts.FindTexts([]string{".*聊聊天.*"}, WithRegex(true))
-	if err == nil {
-		point := points[0].Center()
-		enterPoint = PointF{
-			X: point.X,
-			Y: point.Y - 300,
-		}
-		return enterPoint, true
-	}
-
-	// TODO: 头像入口
-
-	return PointF{}, false
-}
-
-// run live video crawler
-func (vc *VideoCrawler) startLiveCrawler(enterPoint PointF) error {
-	log.Info().Msg("enter live room")
-	if err := vc.driverExt.TapAbsXY(enterPoint.X, enterPoint.Y); err != nil {
-		log.Error().Err(err).Msg("tap live video failed")
-		return err
-	}
-	time.Sleep(5 * time.Second)
-	for !vc.isLiveTargetAchieved() {
-		select {
-		case <-vc.timer.C:
-			log.Warn().Msg("timeout in live crawler")
-			return errors.Wrap(code.TimeoutError, "live crawler timeout")
-		case <-vc.driverExt.interruptSignal:
-			log.Warn().Msg("interrupted in live crawler")
-			return errors.Wrap(code.InterruptError, "live crawler interrupted")
-		default:
-			// swipe to next live video
-			swipeStartTime := time.Now()
-			if err := vc.driverExt.SwipeUp(); err != nil {
-				log.Error().Err(err).Msg("live swipe up failed")
-				return err
-			}
-			swipeFinishTime := time.Now()
-
-			// wait for live video loading
-			time.Sleep(5 * time.Second)
-
-			// TODO: get app event trackings
-			liveRoom, err := vc.getCurrentLiveRoom()
-			if err != nil {
-				return errors.Wrap(err, "get current live event trackings failed")
-			}
-
-			// take screenshot and get screen texts by OCR
-			screenResult, err := vc.driverExt.GetScreenResult(
-				WithScreenShotOCR(true), WithScreenShotUpload(true))
-			if err != nil {
-				log.Error().Err(err).Msg("OCR GetTexts failed")
-				time.Sleep(3 * time.Second)
-				continue
-			}
-			screenResult.Live = liveRoom
-
-			// TODO: check live type
-
-			// incr live count
-			screenResult.VideoType = "live"
-			vc.LiveCount++
-			log.Info().Strs("tags", screenResult.Tags).
-				Interface("live", screenResult.Live).
-				Msg("found live success")
-
-			// get simulation watch duration
-			if screenResult.Live.SimulationWatchDuration != 0 {
-				screenResult.Live.WatchDuration = screenResult.Live.SimulationWatchDuration
-			} else {
-				screenResult.Live.RandomWatchDuration = getSimulationDuration(vc.configs.Live.SleepRandom)
-				screenResult.Live.WatchDuration = screenResult.Live.RandomWatchDuration
-			}
-			// simulation watch live video
-			sleepStrict(swipeFinishTime, screenResult.Live.WatchDuration)
-
-			// log swipe timelines
-			screenResult.SwipeStartTime = swipeStartTime.UnixMilli()
-			screenResult.SwipeFinishTime = swipeFinishTime.UnixMilli()
-			screenResult.TotalElapsed = time.Since(swipeFinishTime).Milliseconds()
-		}
-	}
-
-	log.Info().Msg("live count achieved, exit live room")
-
-	return vc.exitLiveRoom()
-}
-
 func (vc *VideoCrawler) exitLiveRoom() error {
-	for i := 0; i < 3; i++ {
-		vc.driverExt.SwipeRelative(0.1, 0.5, 0.9, 0.5)
-		time.Sleep(2 * time.Second)
-	}
-
-	// exit live room failed, while video count achieved
-	if vc.isTargetAchieved() {
-		return nil
-	}
-
-	// click X button on upper-right corner
-	if err := vc.driverExt.TapXY(0.95, 0.05); err == nil {
-		log.Info().Msg("tap X button on upper-right corner to exit live room")
-		time.Sleep(2 * time.Second)
-	}
-
-	return errors.New("exit live room failed")
+	log.Info().Msg("press back to exit live room")
+	return vc.driverExt.Driver.PressBack()
 }
+
+const (
+	FOUND_FEED_SUCCESS = "found feed success"
+	FOUND_LIVE_SUCCESS = "found live success"
+)
 
 func (dExt *DriverExt) VideoCrawler(configs *VideoCrawlerConfigs) (err error) {
 	if dExt.plugin == nil {
@@ -267,13 +140,10 @@ func (dExt *DriverExt) VideoCrawler(configs *VideoCrawlerConfigs) (err error) {
 		configs:   configs,
 
 		failedCount: 0,
-		lastFeed:    &FeedVideo{},
-		lastLive:    &LiveRoom{},
-
-		FeedCount: 0,
-		FeedStat:  make(map[string]int),
-		LiveCount: 0,
-		LiveStat:  make(map[string]int),
+		FeedCount:   0,
+		FeedStat:    make(map[string]int),
+		LiveCount:   0,
+		LiveStat:    make(map[string]int),
 	}
 	defer func() {
 		dExt.cacheStepData.videoCrawler = crawler
@@ -291,78 +161,152 @@ func (dExt *DriverExt) VideoCrawler(configs *VideoCrawlerConfigs) (err error) {
 			log.Warn().Msg("interrupted in feed crawler")
 			return errors.Wrap(code.InterruptError, "feed crawler interrupted")
 		default:
+			if err = crawler.clearCurrentVideo(); err != nil {
+				log.Error().Err(err).Msg("clear cache failed")
+			}
+
 			// swipe to next feed video
 			log.Info().Msg("swipe to next feed video")
 			swipeStartTime := time.Now()
-			if err = dExt.SwipeUp(); err != nil {
+			if err = dExt.SwipeRelative(0.9, 0.8, 0.9, 0.1, WithOffsetRandomRange(-10, 10)); err != nil {
 				log.Error().Err(err).Msg("feed swipe up failed")
 				return err
 			}
 			swipeFinishTime := time.Now()
 
 			// get app event trackings
-			// retry 3 times if get feed failed, abort if fail 3 consecutive times
-			feedVideo, err := crawler.getCurrentFeedVideo()
-			if err != nil || feedVideo.VideoID == "" {
-				if crawler.failedCount >= 3 {
-					// failed 3 consecutive times
-					return errors.New("get current feed video failed 3 consecutive times")
-				}
-				log.Warn().Interface("feedVideo", feedVideo).Msg("get current feed video failed")
-				// retry
+			// retry 10 times if get feed failed, abort if fail 10 consecutive times
+			currentVideo, err := crawler.getCurrentVideo()
+			if err != nil || currentVideo.Type == "" {
 				crawler.failedCount++
+				if crawler.failedCount >= 10 {
+					// failed 10 consecutive times
+					return errors.Wrap(code.TrackingGetError,
+						"get current feed video failed 10 consecutive times")
+				}
+				log.Warn().
+					Int64("failedCount", crawler.failedCount).
+					Msg("get current feed video failed")
+
+				// check and handle popups
+				if err := crawler.driverExt.ClosePopupsHandler(); err != nil {
+					return err
+				}
+
+				// retry
 				continue
 			}
 
-			if feedVideo.VideoID == crawler.lastFeed.VideoID {
-				// app event tracking not changed
-				// check and handle popups
-				if err = crawler.driverExt.ClosePopupsHandler(WithMaxRetryTimes(1)); err != nil {
-					return err
-				}
-			}
-			crawler.lastFeed = feedVideo
-
-			screenResult := &ScreenResult{}
-			dExt.cacheStepData.screenResults[time.Now().String()] = screenResult
-
-			// check if live video && run live crawler
-			if enterPoint, isLive := crawler.checkLiveVideo(feedVideo); isLive {
+			switch currentVideo.Type {
+			case VideoType_PreviewLive:
 				// 直播预览流
-				screenResult.VideoType = "live-preview"
-				// TODO
-				// screenResult.Live = feedVideo
-				log.Info().Msg("live video found")
-				if !crawler.isLiveTargetAchieved() {
-					if err := crawler.startLiveCrawler(enterPoint); err != nil {
+				var skipEnterLive bool
+				if crawler.isLiveTargetAchieved() {
+					log.Info().Interface("video", currentVideo).
+						Msg("live count achieved, skip entering live room")
+					skipEnterLive = true
+				} else if rand.Float64() <= 0.10 {
+					// 10% chance skip entering live room
+					log.Info().Msg("skip entering preview live by 10% chance")
+					skipEnterLive = true
+				}
+
+				if !skipEnterLive {
+					time.Sleep(1 * time.Second)
+					// enter live room
+					entryPoint := PointF{
+						X: float64(dExt.windowSize.Width / 2),
+						Y: float64(dExt.windowSize.Height / 2),
+					}
+
+					log.Info().Msg("tap screen center to enter live room")
+					if err := crawler.driverExt.TapAbsXY(entryPoint.X, entryPoint.Y,
+						WithOffsetRandomRange(-20, 20)); err != nil {
+						log.Error().Err(err).Msg("tap live video failed")
+						continue
+					}
+				} else {
+					// skip entering live room
+					// only mock simulation play duration
+					sleepTime := math.Min(float64(currentVideo.SimulationPlayDuration), float64(currentVideo.RandomPlayDuration))
+					currentVideo.PlayDuration = int64(sleepTime)
+				}
+
+				fallthrough
+
+			case VideoType_Live:
+				// 直播
+				crawler.LiveCount++
+				log.Info().Interface("video", currentVideo).Msg(FOUND_LIVE_SUCCESS)
+
+				// wait 3s for live loading
+				time.Sleep(3 * time.Second)
+				// take screenshot and get screen texts by OCR
+				screenResult, err := crawler.driverExt.GetScreenResult(
+					WithScreenShotOCR(true),
+					WithScreenShotUpload(true),
+					WithScreenShotLiveType(true),
+				)
+				if err != nil {
+					log.Error().Err(err).Msg("get screen result failed")
+					time.Sleep(3 * time.Second)
+					continue
+				}
+
+				// add live type
+				if screenResult.imageResult != nil &&
+					screenResult.imageResult.LiveType != "" &&
+					screenResult.imageResult.LiveType != "NoLive" {
+					currentVideo.LiveType = screenResult.imageResult.LiveType
+				}
+
+				// simulation watch feed video
+				sleepStrict(swipeFinishTime, currentVideo.PlayDuration)
+
+				screenResult.Video = currentVideo
+				screenResult.Resolution = dExt.windowSize
+				screenResult.SwipeStartTime = swipeStartTime.UnixMilli()
+				screenResult.SwipeFinishTime = swipeFinishTime.UnixMilli()
+				screenResult.TotalElapsed = time.Since(swipeFinishTime).Milliseconds()
+
+				var exitLive bool
+				if crawler.isLiveTargetAchieved() {
+					log.Info().Interface("live", currentVideo).
+						Msg("live count achieved, exit live room")
+					exitLive = true
+				} else if rand.Float64() <= 0.10 {
+					// 10% chance exit live room
+					log.Info().Msg("exit live room by 10% chance")
+					exitLive = true
+				}
+				if exitLive && currentVideo.Type == VideoType_Live {
+					err = crawler.exitLiveRoom()
+					if err != nil {
 						if errors.Is(err, code.TimeoutError) || errors.Is(err, code.InterruptError) {
 							return err
 						}
 						log.Error().Err(err).Msg("run live crawler failed, continue")
-						continue
 					}
 				}
-			} else {
-				// 点播
-				// check feed type and incr feed count
-				screenResult.VideoType = "feed"
-				screenResult.Feed = feedVideo
-				crawler.FeedCount++
-				log.Info().
-					Strs("tags", screenResult.Tags).
-					Interface("feed", screenResult.Feed).
-					Msg("found feed success")
 
-				// get simulation play duration
-				if screenResult.Feed.SimulationPlayDuration != 0 {
-					screenResult.Feed.PlayDuration = screenResult.Feed.SimulationPlayDuration
-				} else {
-					screenResult.Feed.RandomPlayDuration = getSimulationDuration(crawler.configs.Feed.SleepRandom)
-					screenResult.Feed.PlayDuration = screenResult.Feed.RandomPlayDuration
+			default:
+				// 点播 || 图文 || 广告 || etc.
+				crawler.FeedCount++
+				log.Info().Interface("video", currentVideo).Msg(FOUND_FEED_SUCCESS)
+
+				screenResult := &ScreenResult{
+					Resolution: dExt.windowSize,
+					Video:      currentVideo,
+
+					// log swipe timelines
+					SwipeStartTime:  swipeStartTime.UnixMilli(),
+					SwipeFinishTime: swipeFinishTime.UnixMilli(),
 				}
+				dExt.cacheStepData.screenResults[time.Now().String()] = screenResult
 
 				// simulation watch feed video
-				sleepStrict(swipeFinishTime, screenResult.Feed.PlayDuration)
+				sleepStrict(swipeFinishTime, currentVideo.PlayDuration)
+				screenResult.TotalElapsed = time.Since(swipeFinishTime).Milliseconds()
 			}
 
 			// check if target count achieved
@@ -371,96 +315,117 @@ func (dExt *DriverExt) VideoCrawler(configs *VideoCrawlerConfigs) (err error) {
 				return nil
 			}
 
-			// log swipe timelines
-			screenResult.SwipeStartTime = swipeStartTime.UnixMilli()
-			screenResult.SwipeFinishTime = swipeFinishTime.UnixMilli()
-			screenResult.TotalElapsed = time.Since(swipeFinishTime).Milliseconds()
-
 			// reset failed count
 			crawler.failedCount = 0
 		}
 	}
 }
 
-type FeedVideo struct {
-	// 视频基础数据
-	VideoID  string `json:"video_id"`  // 视频 video ID
-	UserName string `json:"user_name"` // 视频作者
-	Duration int64  `json:"duration"`  // 视频时长(ms)
-	Caption  string `json:"caption"`   // 视频文案
-	Type     string `json:"type"`      // 视频类型, feed/live // TODO: 区分视频、图文、广告
+type VideoType string
 
+const (
+	VideoType_Feed        VideoType = "FEED"
+	VideoType_PreviewLive VideoType = "PREVIEW-LIVE" // 直播预览流
+	VideoType_Live        VideoType = "LIVE"
+	VideoType_Image       VideoType = "IMAGE"
+)
+
+type Video struct {
+	Type     VideoType `json:"type" required:"true"` // 视频类型, feed/preview-live/live/image
+	DataType string    `json:"data_type"`            // 数据源对应的事件名称
+
+	// Feed 视频基础数据
+	CacheKey string `json:"cache_key,omitempty"` // cachekey
+	VideoID  string `json:"video_id,omitempty"`  // 视频 video ID
+	URL      string `json:"feed_url,omitempty"`  // 实际播放的视频 url
+	UserName string `json:"user_name"`           // 视频作者
+	Duration int64  `json:"duration,omitempty"`  // 视频时长(ms)
+	Caption  string `json:"caption,omitempty"`   // 视频文案
 	// 视频热度数据
-	ViewCount    int64 `json:"view_count"`    // feed 观看数
-	LikeCount    int64 `json:"like_count"`    // feed 点赞数
-	CommentCount int64 `json:"comment_count"` // feed 评论数
-	CollectCount int64 `json:"collect_count"` // feed 收藏数
-	ForwardCount int64 `json:"forward_count"` // feed 转发数
-	ShareCount   int64 `json:"share_count"`   // feed 分享数
+	ViewCount    int64 `json:"view_count,omitempty"`    // feed 观看数
+	LikeCount    int64 `json:"like_count,omitempty"`    // feed 点赞数
+	CommentCount int64 `json:"comment_count,omitempty"` // feed 评论数
+	CollectCount int64 `json:"collect_count,omitempty"` // feed 收藏数
+	ForwardCount int64 `json:"forward_count,omitempty"` // feed 转发数
+	ShareCount   int64 `json:"share_count,omitempty"`   // feed 分享数
+
+	// timelines
+	PublishTimestamp int64 `json:"publish_timestamp,omitempty"` // feed 发布时间戳
+	PreloadTimestamp int64 `json:"preload_timestamp,omitempty"` // feed 预加载时间戳
+
+	// Live 视频基础数据
+	LiveStreamID  string `json:"live_stream_id,omitempty"`  // 直播流 ID
+	LiveStreamURL string `json:"live_stream_url,omitempty"` // 直播流地址
+	LiveType      string `json:"live_type,omitempty"`       // 直播间类型
+	// 网络数据
+	ThroughputKbps int64 `json:"throughput_kbps,omitempty"` // 网速
+	// 视频热度数据
+	AudienceCount int64 `json:"audience_count,omitempty"` // 直播间人数
+
+	// 图文数据
+	ImageUrls []string `json:"image_urls,omitempty"` // 图片对应的 url 列表
 
 	// 记录仿真决策信息
 	PlayDuration           int64   `json:"play_duration"`            // 播放时长(ms)，取自 Simulation/Random
 	SimulationPlayProgress float64 `json:"simulation_play_progress"` // 仿真播放比例（完播率）
 	SimulationPlayDuration int64   `json:"simulation_play_duration"` // 仿真播放时长(ms)
 	RandomPlayDuration     int64   `json:"random_play_duration"`     // 随机播放时长(ms)
-
-	// timelines
-	PublishTimestamp int64 `json:"publish_timestamp"` // feed 发布时间戳
-	PreloadTimestamp int64 `json:"preload_timestamp"` // feed 预加载时间戳
 }
 
-type LiveRoom struct {
-	// 视频基础数据
-	LiveStreamID string `json:"live_stream_id"` // 直播流 ID
-	UserName     string `json:"user_name"`      // 视频作者
-	Caption      string `json:"caption"`        // 视频文案
-	LiveType     string `json:"live_type"`      // 直播间类型, 基于算法服务获取
-
-	// 视频热度数据
-	AudienceCount string `json:"audience_count"` // 直播间人数
-	LikeCount     int64  `json:"like_count"`     // 点赞数
-
-	// 记录仿真决策信息
-	WatchDuration           int64 `json:"watch_duration"`            // 观播时长(ms)，取自 Simulation/Random
-	SimulationWatchDuration int64 `json:"simulation_watch_duration"` // 仿真观播时长(ms)
-	RandomWatchDuration     int64 `json:"random_watch_duration"`     // 随机观播时长(ms)
-
-	// timelines
-	PreloadTimestamp int64 `json:"preload_timestamp"` // feed 预加载时间戳
-}
-
-func (vc *VideoCrawler) getCurrentFeedVideo() (feedVideo *FeedVideo, err error) {
-	if !vc.driverExt.plugin.Has("GetCurrentFeedVideo") {
-		return nil, errors.New("plugin missing GetCurrentFeedVideo method")
+func (vc *VideoCrawler) clearCurrentVideo() error {
+	if !vc.driverExt.plugin.Has("ClearCurrentVideo") {
+		return errors.New("plugin missing ClearCurrentVideo method")
 	}
 
-	resp, err := vc.driverExt.plugin.Call("GetCurrentFeedVideo")
+	_, err := vc.driverExt.plugin.Call("ClearCurrentVideo")
 	if err != nil {
-		return nil, errors.Wrap(err, "call plugin GetCurrentFeedVideo failed")
+		return errors.Wrap(err, "call plugin ClearCurrentVideo failed")
+	}
+
+	return nil
+}
+
+func (vc *VideoCrawler) getCurrentVideo() (video *Video, err error) {
+	if !vc.driverExt.plugin.Has("GetCurrentVideo") {
+		return nil, errors.New("plugin missing GetCurrentVideo method")
+	}
+
+	resp, err := vc.driverExt.plugin.Call("GetCurrentVideo")
+	if err != nil {
+		return nil, errors.Wrap(err, "call plugin GetCurrentVideo failed")
 	}
 
 	if resp == nil {
-		return nil, errors.New("feed not found")
+		return nil, errors.New("video not found")
 	}
 
 	feedBytes, err := json.Marshal(resp)
 	if err != nil {
-		return nil, errors.New("json marshal feed video info failed")
+		return nil, errors.New("json marshal video info failed")
 	}
 
-	feedVideo = &FeedVideo{}
-	err = json.Unmarshal(feedBytes, feedVideo)
+	video = &Video{}
+	err = json.Unmarshal(feedBytes, video)
 	if err != nil {
-		return nil, errors.Wrap(err, "json unmarshal feed video info failed")
+		return nil, errors.Wrap(err, "json unmarshal video info failed")
+	}
+
+	if video.Type == VideoType_Live || video.Type == VideoType_PreviewLive {
+		video.RandomPlayDuration = getSimulationDuration(vc.configs.Live.SleepRandom)
+	} else {
+		video.RandomPlayDuration = getSimulationDuration(vc.configs.Feed.SleepRandom)
+	}
+
+	// get simulation play duration
+	if video.SimulationPlayDuration != 0 {
+		video.PlayDuration = video.SimulationPlayDuration
+	} else {
+		video.PlayDuration = video.RandomPlayDuration
 	}
 
 	log.Info().
-		Interface("feedVideoCaption", feedVideo.Caption).
-		Msg("get current feed video success")
-	return feedVideo, nil
-}
-
-func (vc *VideoCrawler) getCurrentLiveRoom() (liveVideo *LiveRoom, err error) {
-	// TODO
-	return
+		Str("type", string(video.Type)).
+		Str("dataType", video.DataType).
+		Msg("get current video success")
+	return video, nil
 }